Finish with fake fill db
